TechManager compatibility file for 0.90
Nobody ever¹ has had any problems with TechManager 1.5 for 0.90, and
the author hasn't been seen in WEEKS. Since evrything good in the world²
uses TechManager and the Community Tech Tree, here's a compatibility
shim that will install v1.5 for 0.90.

"Why not change the existing metadata?", I hear you ask. Mainly it's
because it's generated by netkan, and if KerbalStuff ever gets
updated, or if the metadata ever gets regenerated, we don't want
anything to break.

Many thanks to Daz for the prompting.

Attentioning @NathanKell, because relevance.

Also, this means I might finally upgrade to KSP 0.90.

¹ Allegedly.
² RP-0