Update and rename ContractConfigurator-GrandTours-1.0.6.ckan to ContractConfigurator-GrandTours-0.1.6.ckan
Looking at the versioning over at KerbalStuffs changelog it would seem rather apperant that this version is an error and it's preventing the latest version (with crucial bugfixes) to reach users.

I'm not sure if this is the correct way of solving this issue or if we should bump epoch? How will this affect users in the downstream with 1.0.6 installed?