Rewrite versions to be more specific!
As discussed in KSP-CKAN/CKAN#1156, sometimes we see older versions of
mods that have "generic" KSP versions (eg: 1.0), even though newer
versions of those same mods have specific versions (eg 1.0.1). These has
been a side-effect of how KS returns KSP version numbers to us.

This patch adds ".0" to version numbers if and only if we see a later
version of the same mod that has a specific version number in the same
sequence (eg: 1.0 and 1.0.3 causes a rewrite, but 0.24 and 1.0.3 does
not).

Source code coming soon. I haven't checked it into CKAN-meta because
it's not really a place where lots of code should be hanging out.

Also attn @dbent, who identified this issue in the first place.