Patial recovery of KS mods.
These mods are:

1. All under FOSS licenses.
1. All present in our indexer, and so easily available for me to upload.
1. Now in a great big S3 bucket.

S3 buckets have interfaces that make me cry, and there's nowhere near as
many of these as I would like, but it's a start.

The S3 buckets I'm using will run out in a couple of years, so
longer-term we should migrate these to the Internet Archive. Yes,
they're cool with this (provided all the content is under a freely
distributable license), but it requires some fussing to get up and
running. So I've gone with S3 for now.

I've tested some, but not all of these. You can test them by going to
CKAN -> Settings and using a metadata repo of
`https://github.com/KSP-CKAN/CKAN-meta/archive/ks_recovery_1.tar.gz`.

Obviously only the migrated mods (see filelist) will work with this
migration. But it's a start.