RN fix my mods that are set for all ksp versions
Somehow these 3 releases gt set for "all ksp versions", this is
extremely incorrect as they absolutely do NOT work in all versions. This
is making ckan install the older releases in 1.4.2 even though the
latest release only supports up to 1.3.1.

-These mods DID contain version files when these meta data files were
generated and for some reason ignored my versioning in them