Update Kerbalism-1.4.1.4.ckan
For some reason this metadata is incorrect. I have inflated this version locally using the latest NetKAN and it works for me while the one currently in CKAN-meta does not. I'm not sure how we fix these things anymore so I'm creating this PR in hopes of someone more knowledgable telling me whether this is the correct way or not. If netkan-bot overwrites this on the next run and puts back the incorrect values I guess we'll need to take another look.