Remove deleted release
I have deleted this release from the github repo becasue it shouldn't
have been released before the 1.4.5 version for ksp. I am going to make
a new release for ksp 1.4.5 then remake this release for 1.5.1 after
that. This file will be recreated then and this one will need to be
deleted as it's non functional now