Update SimpleConstruction-1-4.0.2.6.ckan
Was processing a cry for help on the forums when I saw the 'Updated' tag. Still the same mod so don't need that. :D adopted with full permission and blessing of original mod author.
📦 〽️ 

also - thought I would update the listings - haven't a clue what 'DeadskinsMeshSwitch' is still doing there. :D